# CUDA: fix tensor core logic for Pascal and HIP