https://github.com/ggerganov/llama.cpp/pull/3776 has made it so that FP16 cuBLAS GEMM is used instead of MMQ for Volta or newer. However, the way the logic was implemented is incorrect for Pascal and HIP. With Pascal `use_tensor_cores` is evaluated as `true` which results in FP16 cuBLAS GEMM being used even though Pascal has terrible FP16 performance. For AMD GPUs hipBLAS also seems to perform worse than MMQ. This PR makes it so that `use_tensor_cores` is evaluated as `false` if the minimum compute capability is lower than Volta or when using HIP.

| GPU     | Model   | Test  | t/s master | t/s PR | Speedup |
|---------|---------|-------|------------|--------|---------|
| P40     | 7b q8_0 | pp512 |        334 |    894 |    2.68 |
| P40     | 7b q8_0 | tg128 |      24.68 |  34.63 |     1.4 |
| RX 6800 | 7b q8_0 | pp512 |        660 |   1276 |    1.93 |
| RX 6800 | 7b q8_0 | tg128 |      49.16 |  47.98 |    0.98 |