@slaren 2023-12-29
Doesn't AMD have an equivalent of tensor cores with WMMA? Aren't any cases where using F16 mat mul with hipBLAS is beneficial for AMD hardware? If I understand correctly, it is available on RDNA3 hardware.