@sorasoras 2023-12-29
> I unfortunately don't have an RX 7000 GPU to test and AMD does not report tensor core FLOPS either. But with my RTX 3090 I get ~1.5 better speeds with FP16 tensor cores compared to MMQ even though the FLOPS number that NVIDIA provides is ~5x higher. So I don't think there will be a large decline in RX 7000 performance (if there is I will make a follow-up PR).

For RX 7000， prompt eval is speed is half but total gen speed is not huge different. MMQ give you less latency but tensor is faster with longer context.