@JohannesGaessler 2023-12-29
I expanded the logic a bit. there are now two booleans: `fp16_performance_good` and `use_mul_mat_q`. This allows you to select FP16 BLAS and MMQ independently of one another. For my RTX 3090 and my RX 6800 using FP16 BLAS is faster regardless of whether or not MMQ is used for the quantized matrix multiplication. @sorasoras please re-test this PR, there should no longer be a performance regression.