@JohannesGaessler 2023-12-29
Since there has been no further feedback so far I've decided to just merge the PR. IF there are still issues I'll make a new one.