@sorasoras 2023-12-30
> Since there has been no further feedback so far I've decided to just merge the PR. If there are still issues I'll make a new one.
branch 1728
print_timings: prompt eval time =     472.72 ms /   409 tokens (    1.16 ms per token,   865.20 tokens per second)
print_timings:        eval time =    4836.31 ms /   246 runs   (   19.66 ms per token,    50.87 tokens per second)
print_timings:       total time =    5309.03 ms
slot 0 released (656 tokens in cache)
{"timestamp":1703909975,"level":"INFO","function":"log_server_request","line":2677,"message":"request","remote_addr":"127.0.0.1","remote_port":60816,"status":200,"method":"POST","path":"/completion","params":{}}
slot 1 : kv cache rm - [0, end)
{"timestamp":1703909975,"level":"INFO","function":"log_server_request","line":2677,"message":"request","remote_addr":"127.0.0.1","remote_port":60816,"status":200,"method":"OPTIONS","path":"/completion","params":{}}
slot 0 is processing [task id: 26]

branch1727

slot 1 is processing [task id: 5]

print_timings: prompt eval time =     377.81 ms /   454 tokens (    0.83 ms per token,  1201.67 tokens per second)
print_timings:        eval time =    5016.92 ms /   253 runs   (   19.83 ms per token,    50.43 tokens per second)
print_timings:       total time =    5394.73 ms
slot 0 released (708 tokens in cache)
{"timestamp":1703910621,"level":"INFO","function":"log_server_request","line":2677,"message":"request","remote_addr":"127.0.0.1","remote_port":64464,"status":200,"method":"POST","path":"/completion","params":{}}
slot 1 : kv cache rm - [0, end)
{"timestamp":1703910621,"level":"INFO","function":"log_server_request","line":2677,"message":"request","remote_addr":"127.0.0.1","remote_port":64491,"status":200,"method":"OPTIONS","path":"/completion","params":{}}
slot 0 is processing [task id: 6]


The bigger difference is the prompt eval time for longer input.
other than that , they are the same.