@JohannesGaessler 2023-12-30
What does the console say on branch 1728 for `USE_TENSOR_CORES`? It may be that I incorrectly assumed that the macros for gfx versions would be defined in CPU code.