@sorasoras 2023-12-30
> What does the console say on branch 1728 for `USE_TENSOR_CORES`? It may be that I incorrectly assumed that the macros for gfx versions would be defined in CPU code.

ggml_init_cublas: GGML_CUDA_FORCE_MMQ:   no
ggml_init_cublas: CUDA_USE_TENSOR_CORES: no