@DaveYognaught 2024-02-05
Unfortunately; This change has resulted in extremely poor performance, in the 3x range on average, for the entire NVIDIA GTX 1600 series. Is there a better way of calculating this? The compute capability of our GPUs is set at 7.5, but we likewise do not tensor cores and the FP16 performance is awful. 

Before:  32.7ms/T to 54.5ms/T (AVG: 44ms/T)
After:    64.6ms/T to 224ms/T  (AVG: 131.35ms/T)