@Artem-B 2024-02-05
Interesting. Looks like all TU116 and TU117 GPU variants do not have tensorcores, despite being nominally sm_75: https://www.techpowerup.com/gpu-specs/?architecture=Turing&sort=generation